Progress on Fairness System
Some nasty bug i can't find right now. Every GetGridCollision call
returns 0 EXCEPT POOP!!! WHAAAT??
(Pressing 3 ingame will switch to grid-fairness-debug-mode and will show
the fairness value of every gridtile.
The fairness-value is calculated by the amount of free space around the
tile.
For ex: A tile touching a wall will always have a fairness of 3 or
lower.
A tile touching two walls will always have a fairness of 2 or lower.
A tile which is completetly surrounded by air always has a fairness of
4.

Rocks and other grid-tiles which are completly solid should not contain
a value. But for some reason only poop is recognized as solid. help)